lara/common: port Lara_UseItem to TRX
