+++ src/libtrx/include/libtrx/game/lara/common.h
@@ -10,6 +10,7 @@ ITEM *Lara_GetItem(void);
 void Lara_Initialise(const GF_LEVEL *level);
 void Lara_InitialiseInventory(const GF_LEVEL *level);
 void Lara_RevertToPistolsIfNeeded(void);
+void Lara_UseItem(GAME_OBJECT_ID obj_id);
 void Lara_SetStartAnimState(LARA_EXTRA_STATE state);
 bool Lara_IsControllable(void);
 void Lara_SetControllable(bool controllable);
