+++ src/libtrx/include/libtrx/game/stats/common.h
@@ -14,5 +14,6 @@ extern uint32_t Stats_GetMaxSecretFlags(void);
 
 void Stats_UpdateSecrets(LEVEL_STATS *stats);
 
+void Stats_AddMedipacksUsed(double medipack_value);
 void Stats_AddDeath(void);
 void Stats_AddDistanceTravelled(XYZ_32 pos, XYZ_32 last_pos);
