+++ src/tr1/game/lara/common.c
@@ -31,114 +31,6 @@ ITEM *Lara_GetItem(void)
     return g_LaraItem;
 }
 
-void Lara_UseItem(const GAME_OBJECT_ID obj_id)
-{
-    LOG_INFO("%d", obj_id);
-    switch (obj_id) {
-    case O_PISTOL_ITEM:
-    case O_PISTOL_OPTION:
-        g_Lara.request_gun_type = LGT_PISTOLS;
-        if (g_Lara.gun_status == LGS_ARMLESS
-            && g_Lara.gun_type == LGT_PISTOLS) {
-            g_Lara.gun_type = LGT_UNARMED;
-        }
-        break;
-
-    case O_SHOTGUN_ITEM:
-    case O_SHOTGUN_OPTION:
-        g_Lara.request_gun_type = LGT_SHOTGUN;
-        if (g_Lara.gun_status == LGS_ARMLESS
-            && g_Lara.gun_type == LGT_SHOTGUN) {
-            g_Lara.gun_type = LGT_UNARMED;
-        }
-        break;
-
-    case O_MAGNUM_ITEM:
-    case O_MAGNUM_OPTION:
-        g_Lara.request_gun_type = LGT_MAGNUMS;
-        if (g_Lara.gun_status == LGS_ARMLESS
-            && g_Lara.gun_type == LGT_MAGNUMS) {
-            g_Lara.gun_type = LGT_UNARMED;
-        }
-        break;
-
-    case O_UZI_ITEM:
-    case O_UZI_OPTION:
-        g_Lara.request_gun_type = LGT_UZIS;
-        if (g_Lara.gun_status == LGS_ARMLESS && g_Lara.gun_type == LGT_UZIS) {
-            g_Lara.gun_type = LGT_UNARMED;
-        }
-        break;
-
-    case O_SMALL_MEDIPACK_ITEM:
-    case O_SMALL_MEDIPACK_OPTION:
-        if (g_LaraItem->hit_points <= 0
-            || g_LaraItem->hit_points >= LARA_MAX_HITPOINTS) {
-            return;
-        }
-        g_LaraItem->hit_points += LARA_MAX_HITPOINTS / 2;
-        CLAMPG(g_LaraItem->hit_points, LARA_MAX_HITPOINTS);
-        Inv_RemoveItem(O_SMALL_MEDIPACK_ITEM);
-        Sound_Effect(SFX_MENU_MEDI, nullptr, SPM_ALWAYS);
-        Stats_AddMedipacksUsed(.5);
-        break;
-
-    case O_LARGE_MEDIPACK_ITEM:
-    case O_LARGE_MEDIPACK_OPTION:
-        if (g_LaraItem->hit_points <= 0
-            || g_LaraItem->hit_points >= LARA_MAX_HITPOINTS) {
-            return;
-        }
-        g_LaraItem->hit_points = g_LaraItem->hit_points + LARA_MAX_HITPOINTS;
-        CLAMPG(g_LaraItem->hit_points, LARA_MAX_HITPOINTS);
-        Inv_RemoveItem(O_LARGE_MEDIPACK_ITEM);
-        Sound_Effect(SFX_MENU_MEDI, nullptr, SPM_ALWAYS);
-        Stats_AddMedipacksUsed(1);
-        break;
-
-    case O_KEY_ITEM_1:
-    case O_KEY_OPTION_1:
-    case O_KEY_ITEM_2:
-    case O_KEY_OPTION_2:
-    case O_KEY_ITEM_3:
-    case O_KEY_OPTION_3:
-    case O_KEY_ITEM_4:
-    case O_KEY_OPTION_4:
-    case O_PUZZLE_ITEM_1:
-    case O_PUZZLE_OPTION_1:
-    case O_PUZZLE_ITEM_2:
-    case O_PUZZLE_OPTION_2:
-    case O_PUZZLE_ITEM_3:
-    case O_PUZZLE_OPTION_3:
-    case O_PUZZLE_ITEM_4:
-    case O_PUZZLE_OPTION_4:
-    case O_LEADBAR_ITEM:
-    case O_LEADBAR_OPTION:
-    case O_SCION_ITEM_1:
-    case O_SCION_ITEM_2:
-    case O_SCION_ITEM_3:
-    case O_SCION_ITEM_4:
-    case O_SCION_OPTION: {
-        int16_t receptacle_item_num = Object_FindReceptacle(obj_id);
-        if (receptacle_item_num == NO_ITEM) {
-            Sound_Effect(SFX_LARA_NO, nullptr, SPM_NORMAL);
-            return;
-        }
-        if (g_Lara.interact_target.item_num != NO_ITEM) {
-            Sound_Effect(SFX_LARA_NO, nullptr, SPM_NORMAL);
-            return;
-        }
-        g_Lara.interact_target.item_num = receptacle_item_num;
-        g_Lara.interact_target.is_moving = true;
-        g_Lara.interact_target.move_count = 0;
-        break;
-    }
-
-    default:
-        break;
-    }
-}
-
 void Lara_InitialiseLoad(int16_t item_num)
 {
     g_Lara.item_num = item_num;
