+++ src/tr1/game/lara/common.h
@@ -7,5 +7,3 @@
 #include <libtrx/game/lara/common.h>
 
 void Lara_InitialiseLoad(int16_t item_num);
-
-void Lara_UseItem(GAME_OBJECT_ID obj_id);
