+++ src/tr2/game/lara/control.c
@@ -11,108 +11,6 @@
 #include <libtrx/game/camera.h>
 #include <libtrx/game/lara.h>
 
-void Lara_UseItem(const GAME_OBJECT_ID obj_id)
-{
-    ITEM *const item = g_LaraItem;
-
-    switch (obj_id) {
-    case O_PISTOL_ITEM:
-    case O_PISTOL_OPTION:
-        g_Lara.request_gun_type = LGT_PISTOLS;
-        break;
-
-    case O_SHOTGUN_ITEM:
-    case O_SHOTGUN_OPTION:
-        g_Lara.request_gun_type = LGT_SHOTGUN;
-        break;
-
-    case O_MAGNUM_ITEM:
-    case O_MAGNUM_OPTION:
-        g_Lara.request_gun_type = LGT_MAGNUMS;
-        break;
-
-    case O_UZI_ITEM:
-    case O_UZI_OPTION:
-        g_Lara.request_gun_type = LGT_UZIS;
-        break;
-
-    case O_HARPOON_ITEM:
-    case O_HARPOON_OPTION:
-        g_Lara.request_gun_type = LGT_HARPOON;
-        break;
-
-    case O_M16_ITEM:
-    case O_M16_OPTION:
-        g_Lara.request_gun_type = LGT_M16;
-        break;
-
-    case O_GRENADE_ITEM:
-    case O_GRENADE_OPTION:
-        g_Lara.request_gun_type = LGT_GRENADE;
-        break;
-
-    case O_SMALL_MEDIPACK_ITEM:
-    case O_SMALL_MEDIPACK_OPTION:
-        if (item->hit_points > 0 && item->hit_points < LARA_MAX_HITPOINTS) {
-            item->hit_points += LARA_MAX_HITPOINTS / 2;
-            CLAMPG(item->hit_points, LARA_MAX_HITPOINTS);
-            Inv_RemoveItem(O_SMALL_MEDIPACK_ITEM);
-            Sound_Effect(SFX_MENU_MEDI, nullptr, SPM_ALWAYS);
-            Stats_AddMedipacksUsed(0.5);
-        }
-        break;
-
-    case O_LARGE_MEDIPACK_ITEM:
-    case O_LARGE_MEDIPACK_OPTION:
-        if (item->hit_points > 0 && item->hit_points < LARA_MAX_HITPOINTS) {
-            item->hit_points = LARA_MAX_HITPOINTS;
-            Inv_RemoveItem(O_LARGE_MEDIPACK_ITEM);
-            Sound_Effect(SFX_MENU_MEDI, nullptr, SPM_ALWAYS);
-            Stats_AddMedipacksUsed(1);
-        }
-        break;
-
-    case O_FLARES_ITEM:
-    case O_FLARES_OPTION:
-        g_Lara.request_gun_type = LGT_FLARE;
-        break;
-
-    case O_KEY_ITEM_1:
-    case O_KEY_OPTION_1:
-    case O_KEY_ITEM_2:
-    case O_KEY_OPTION_2:
-    case O_KEY_ITEM_3:
-    case O_KEY_OPTION_3:
-    case O_KEY_ITEM_4:
-    case O_KEY_OPTION_4:
-    case O_PUZZLE_ITEM_1:
-    case O_PUZZLE_OPTION_1:
-    case O_PUZZLE_ITEM_2:
-    case O_PUZZLE_OPTION_2:
-    case O_PUZZLE_ITEM_3:
-    case O_PUZZLE_OPTION_3:
-    case O_PUZZLE_ITEM_4:
-    case O_PUZZLE_OPTION_4: {
-        const int16_t receptacle_item_num = Object_FindReceptacle(obj_id);
-        if (receptacle_item_num == NO_ITEM) {
-            Sound_Effect(SFX_LARA_NO, nullptr, SPM_NORMAL);
-            return;
-        }
-        if (g_Lara.interact_target.item_num != NO_ITEM) {
-            Sound_Effect(SFX_LARA_NO, nullptr, SPM_NORMAL);
-            return;
-        }
-        g_Lara.interact_target.item_num = receptacle_item_num;
-        g_Lara.interact_target.is_moving = true;
-        g_Lara.interact_target.move_count = 0;
-        break;
-    }
-
-    default:
-        break;
-    }
-}
-
 void Lara_InitialiseLoad(const int16_t item_num)
 {
     g_Lara.item_num = item_num;
