+++ src/tr2/game/lara/control.h
@@ -4,5 +4,4 @@
 
 #include <libtrx/game/lara/common.h>
 
-void Lara_UseItem(GAME_OBJECT_ID obj_id);
 void Lara_InitialiseLoad(int16_t item_num);
